Pi Cam quirk improvements
Added API calls to disable auto-white-balance on exposure set
Revised exposure time to use "raw_" interface
Added scaling logic to remap a 0-100 range exposure command to the 1-10000 range of the Pi V2 Camera's "raw_" interface.
Changed the setExposure() api to take a floating point (not integer) argument

+++ photon-server/src/main/java/org/photonvision/vision/pipeline/CVPipelineSettings.java
@@ -39,7 +39,7 @@ public class CVPipelineSettings implements Cloneable {
     public ImageFlipMode inputImageFlipMode = ImageFlipMode.NONE;
     public ImageRotationMode inputImageRotationMode = ImageRotationMode.DEG_0;
     public String pipelineNickname = "New Pipeline";
-    public int cameraExposure = 50;
+    public double cameraExposure = 50;
     public int cameraBrightness = 50;
     public int cameraGain = 50;
     public int cameraVideoModeIndex = 0;
